+++ warp-drive-packages/core/src/reactive/-private/kind/schema-array-field.ts
@@ -1,22 +1,17 @@
-import type { Store } from '../../../store/-private';
-import type { StableRecordIdentifier } from '../../../types';
-import type { SchemaArrayField } from '../../../types/schema/fields';
+import { assert } from '@warp-drive/build-config/macros';
 
-export function getSchemaArrayField(
-  store: Store,
-  record: object,
-  resourceKey: StableRecordIdentifier,
-  field: SchemaArrayField,
-  path: string | string[],
-  editable: boolean
-): unknown {
-  const { cache } = store;
-  return editable ? cache.getAttr(resourceKey, path) : cache.getRemoteAttr(resourceKey, path);
-}
+import { ARRAY_SIGNAL, type Store } from '../../../store/-private';
+import type { StableRecordIdentifier } from '../../../types';
+import type { ArrayValue } from '../../../types/json/raw';
+import type { SchemaArrayField } from '../../../types/schema/fields';
+import { ManagedArrayMap, peekManagedArray } from '../fields/compute';
+import type { ReactiveResource } from '../record';
+
+export { getArrayField as getSchemaArrayField } from './array-field';
 
 export function setSchemaArrayField(
   store: Store,
-  record: object,
+  record: ReactiveResource,
   resourceKey: StableRecordIdentifier,
   field: SchemaArrayField,
   path: string | string[],
@@ -24,5 +19,19 @@ export function setSchemaArrayField(
 ): boolean {
   const { cache } = store;
 
+  const arrayValue = (value as ArrayValue)?.slice();
+  if (!Array.isArray(arrayValue)) {
+    ManagedArrayMap.delete(record);
+  }
+  cache.setAttr(resourceKey, path, arrayValue);
+  const peeked = peekManagedArray(record, field);
+  if (peeked) {
+    assert(`Expected the peekManagedArray for ${field.kind} to return a ManagedArray`, ARRAY_SIGNAL in peeked);
+    const arrSignal = peeked[ARRAY_SIGNAL];
+    arrSignal.isStale = true;
+  }
+  if (!Array.isArray(value)) {
+    ManagedArrayMap.delete(record);
+  }
   return true;
 }
